Add UTs for new methods

+++ src/NuGetGallery.Services/PackageManagement/PackageOwnershipManagementService.cs
@@ -295,6 +295,11 @@ namespace NuGetGallery
 
         public async Task DeletePackageOwnershipRequestWithMessagesAsync(PackageRegistration packageRegistration, User requestingOwner, User newOwner)
         {
+            if (requestingOwner == null)
+            {
+                throw new ArgumentNullException(nameof(requestingOwner));
+            }
+
             await DeletePackageOwnershipRequestAsync(packageRegistration, newOwner);
 
             var emailMessage = new PackageOwnershipRequestCanceledMessage(_appConfiguration, requestingOwner, newOwner, packageRegistration);
